Created activities to let user to decide data edit format (list vs map)
