+++ app/src/main/java/com/example/recyclerview/MapViewActivity.java
@@ -1,0 +1,14 @@
+package com.example.recyclerview;
+
+import androidx.appcompat.app.AppCompatActivity;
+
+import android.os.Bundle;
+
+public class MapViewActivity extends AppCompatActivity {
+
+    @Override
+    protected void onCreate(Bundle savedInstanceState) {
+        super.onCreate(savedInstanceState);
+        setContentView(R.layout.activity_map_view);
+    }
+}
