+++ app/src/main/res/layout/activity_main.xml
@@ -7,19 +7,27 @@
     android:orientation="vertical"
     tools:context="com.codinginflow.searchfunctionalityexample.MainActivity">
 
-    <EditText
-        android:id="@+id/edittext"
-        android:hint="                          Enter Location"
+    <TextView
+        android:id="@+id/textView3"
+        android:layout_width="match_parent"
+        android:layout_height="236dp"
+        android:text="Lorem ipsum dolor sit amet, consectetur adipiscing elit, sed do eiusmod tempor incididunt ut labore et dolore magna aliqua. Ut enim ad minim veniam, quis nostrud exercitation ullamco laboris nisi ut aliquip ex ea commodo consequat. Duis aute irure dolor in reprehenderit in voluptate velit esse cillum dolore eu fugiat nulla pariatur. Excepteur sint occaecat cupidatat non proident, sunt in culpa qui officia deserunt mollit anim id est laborum." />
+
+    <Button
+        android:id="@+id/noEditButton"
         android:layout_width="match_parent"
         android:layout_height="wrap_content"
-        android:layout_margin="16dp"/>
+        android:text="No. I'm Fine" />
 
-    <androidx.recyclerview.widget.RecyclerView
-        android:id="@+id/recyclerView"
+    <Button
+        android:id="@+id/yesEditListButton"
         android:layout_width="match_parent"
-        android:layout_height="match_parent"
-        android:background="@android:color/darker_gray"
-        android:padding="4dp"
-        android:scrollbars="vertical" />
+        android:layout_height="wrap_content"
+        android:text="Show me data in a list" />
 
+    <Button
+        android:id="@+id/yesEditMapButton"
+        android:layout_width="match_parent"
+        android:layout_height="wrap_content"
+        android:text="Allow me to implement spatial privacy" />
 </LinearLayout>
